Update to use containerised S3 bucket - Can now use a S3 bucket with Docker instead of local file system. - Updated init-s3.sh to ignore directories.

+++ scripts/init-s3.sh
@@ -4,6 +4,8 @@ bucket_name=test-bucket
 awslocal s3 mb s3://$bucket_name || true
 
 DIR=$(realpath ./test_bucket/org/usr/files)
-for file in $DIR/* ; do
-    awslocal s3 cp $file s3://$bucket_name/org/usr/files/$(basename -- $file) ;
+for file in $DIR/*; do
+    if [[ -f $file ]]; then
+        awslocal s3 cp $file s3://$bucket_name/org/usr/files/$(basename -- $file)
+    fi
 done
